examples/python-multi-agent/requirements.txt
openai==1.54.0
python-dotenv==1.0.0



